False parameter not used warning
Issue: A parameter used as part of an expression triggers a false
warning like "The following system parameters are not used...".
Solution: 1) Prepend/append space " " to expression (haystack) and
parameter (needle). 2) Replace operators in haystack by space to avoid
finding parameters which are substrings of other parameters. 3) Search
for needle in haystack.
Before; expression searched for in parameter name so parameter "X"
wouldn't match against for instance "3+X".